+++ client/dashboard/agency/marketplace/exclusive-offers/index.tsx
@@ -1,7 +1,7 @@
 import { __ } from '@wordpress/i18n';
-import { useAnalytics } from '../../app/analytics';
-import { PageHeader } from '../../components/page-header';
-import PageLayout from '../../components/page-layout';
+import { useAnalytics } from '../../../app/analytics';
+import { PageHeader } from '../../../components/page-header';
+import PageLayout from '../../../components/page-layout';
 import PartnerOffers from './partner-offers';
 
 export default function ExclusiveOffers() {
